+++ SpellsPlugin/src/com/elmakers/mine/bukkit/plugins/spells/SpellsPlugin.java
@@ -21,30 +21,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import com.elmakers.mine.bukkit.plugins.groups.Permissions;
 import com.elmakers.mine.bukkit.plugins.groups.PlayerPermissions;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.AbsorbSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.AlterSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.ArrowSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.BlastSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.BlinkSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.BridgeSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.ConstructSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.CushionSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.ExtendSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.FamiliarSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.FillSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.FireballSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.FrostSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.GillsSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.HealSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.MineSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.PillarSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.StairsSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.TorchSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.TowerSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.TransmuteSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.TreeSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.TunnelSpell;
-import com.elmakers.mine.bukkit.plugins.spells.builtin.UndoSpell;
+import com.elmakers.mine.bukkit.plugins.spells.builtin.*;
 import com.elmakers.mine.bukkit.plugins.spells.utilities.BlockList;
 import com.elmakers.mine.bukkit.plugins.spells.utilities.PluginProperties;
 import com.elmakers.mine.bukkit.plugins.spells.utilities.UndoQueue;
@@ -576,6 +553,7 @@ public class SpellsPlugin extends JavaPlugin
 		addSpell(new FamiliarSpell());
 		addSpell(new ConstructSpell());
 		addSpell(new TransmuteSpell());
+		addSpell(new RecallSpell());
 	}
 
 }