+++ SpellsPlugin/src/com/elmakers/mine/bukkit/plugins/spells/builtin/FamiliarSpell.java
@@ -282,7 +282,7 @@ public class FamiliarSpell extends Spell
 	@Override
 	public Material getMaterial()
 	{
-		return Material.EGG;
+		return Material.APPLE;
 	}
 	
 }